Secure the ingresses to fix plex and traefik-forward-auth issues.  (#24)
Description of the change

Plex now works locally and through remote connection. This required TLS to be forwarded from ingress to the backend service to enable secure connection inside plex and to have encryption in place.
Now TLS termination doesn't happen in ingress controller for other services as well and individual certificates are generated automatically for each service with the help of cert-manager and a cluster issuer using Let's Encrypt.
Traefik-forward-auth is now also secure and correctly redirects to https after successful authentication.